Changement du type de variable pour la temperature, l humidite et la pression

+++ src/main/java/toma/meteo/bean/ReleveMeteo.java
@@ -1,5 +1,7 @@
 package toma.meteo.bean;
 
+import java.math.BigDecimal;
+
 import org.springframework.stereotype.Component;
 
 import lombok.AccessLevel;
@@ -17,7 +19,7 @@ import lombok.experimental.FieldDefaults;
 @Component
 public class ReleveMeteo {
 
-	private float temperature;
+	private BigDecimal temperature;
 	private float pression;
 	private float humidite;
